Make PlayerID and Coord usize, set user ID at game reset
Since we're using both a lot for offsets into vectors etc. there is no
reason not to use usize. We're never using them in any other context, basically.

Game reset (GAME) conveys the dimensions of the world as well as our
user ID. Set the user ID. We will set the world in a separate upcoming commit.