End program gracefully by pressing enter
Spawn canary thread that listens on stdin for newline and then tells
the main thread to exit.

This has the side effect that after pressing enter we have to wait for
the next message from the server, e.g. for the next game to begin, in
order for the program to actually exit. To circumvent this we could
set_read_timeout on the stream, but that would have a big drawback: We
cannot join a game that doesn't begin during the first ten seconds
after launching the program.

Also:
- Log amount of empty messages received every time we receive an empty
  message instead of every time we receive a non-empty one.